Update README.md
For Java 1.8, there is 8u211-and-later and 8u202-and-earlier. This changes the link into a direct link to 8u211-and-later, which I assume is the version wanted. Unless 8u202-and-earlier also works and/or there is a reason to not directly go to 1.8 since that is the version of Java explicitly stated to be needed.