Update README.md
Made the verbiage for Maven and Java similar. Mostly reflects original Java verbiage with a small tweak since it reads better due to the differences in how environmental variables are set. Not 100% sold on this exact verbiage, open to suggestions or I'll think on it more. Just wanted consistency established first.